Add test for fethVantageSockData
